requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
phonenumbers==8.12.3
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
